contrib/dimfeld/httptreemux/v5: fix resource renamer
This fixes an issue where the paramater substitution would fail if a
parameter value was contained in the value of another.

For example given the path '/foo/:x/:y/bar' and the request URL
'/foo/123/2/bar', '2' is the value of parameter 'y' but also
contained in the value of parameter 'x', i.e. '123. This could lead
to a scenario where the resource name would be '/foo/1:y3/2/bar'
after the substitution.