contrib/dimfeld/httptreemux.v5: fix resource name for 301 redirects
The httptreemux router has a set of redirect behaviours that is invoked
when a request URL and matched route only differs in a trailing slash.
The default behaviour is to respond with a 301 (moved permanently) to
redirect the client to the exact path of the matched handler. We
previously patched one of the two scenarios where this occurs in #2332.
The changes in this commit addresses the other scenario.

We also standardize the resource name in the event that there is a
trailing slash missmatch between request URL and matched handler. We
always truncate the trailing slash in the resource name.

Fixes #2663